+++ src/server/app.ts
@@ -19,7 +19,7 @@ import { Request, Response, Router, Handler } from 'express';
 import * as hsts from 'hsts';
 
 import * as path from 'path';
-import * as logger from 'morgan';
+import * as morgan from 'morgan';
 import * as bodyParser from 'body-parser';
 import * as compress from 'compression';
 
@@ -31,8 +31,8 @@ if (!WallTime.rules) {
 }
 
 import { GetSettingsOptions } from '../server/utils/settings-manager/settings-manager';
-import { PivotRequest } from './utils/index';
-import { VERSION, AUTH, SERVER_SETTINGS, SETTINGS_MANAGER, LOGGER } from './config';
+import { PivotRequest, LOGGER, TRACKER } from './utils/index';
+import { VERSION, AUTH, SERVER_SETTINGS, SETTINGS_MANAGER } from './config';
 import * as plywoodRoutes from './routes/plywood/plywood';
 import * as plyqlRoutes from './routes/plyql/plyql';
 import * as pivotRoutes from './routes/pivot/pivot';
@@ -84,9 +84,28 @@ function addGuardedRoutes(attach: string, guard: string, router: Router | Handle
   app.use(SERVER_SETTINGS.getServerRoot() + attach, guardHandler, router);
 }
 
+// Add compression
 app.use(compress());
-app.use(logger('dev'));
 
+// Add request logging and tracking
+var morganFormat = SERVER_SETTINGS.getRequestLogFormat();
+var morgenFormatFunction = morgan.compile((morgan as any)[morganFormat] || morganFormat);
+app.use(morgan((m: any, req: PivotRequest, res: Response) => {
+  TRACKER.track({
+    eventType: 'request',
+    attr: {
+      url: m['url'](req, res),
+      method: m['method'](req, res),
+      status: m['status'](req, res)
+    },
+    user: req.user,
+    metric: 'request/time',
+    value: Number(m['response-time'](req, res))
+  });
+  return morgenFormatFunction(m, req, res);
+}));
+
+// Add Strict Transport Security
 if (SERVER_SETTINGS.getStrictTransportSecurity() === "always") {
   app.use(hsts({
     maxAge: 10886400000,     // Must be at least 18 weeks to be approved by Google