+++ src/server/utils/index.ts
@@ -20,3 +20,4 @@ export * from './general/general';
 export * from './settings-manager/settings-manager';
 export * from './updater/updater';
 export * from './logger/logger';
+export * from './tracker/tracker';