+++ test/simulate/reintrospect-on-load-datasource.mocha.js
@@ -23,6 +23,7 @@ const extractConfig = require('../utils/extract-config');
 
 const TEST_PORT = 18082;
 var pivotServer;
+var druidServer;
 
 var segmentMetadataResponse = [
   {
@@ -75,7 +76,7 @@ describe('reintrospect on load with datasource', function () {
   var dataLoaded = false;
 
   before((done) => {
-    mockDruid(28089, {
+    druidServer = mockDruid({
       onDataSources: function() {
         return {
           json: dataLoaded ? ['wikipedia'] : []
@@ -112,10 +113,12 @@ describe('reintrospect on load with datasource', function () {
             throw new Error(`unknown query ${query.queryType}`);
         }
       }
-    }).then(function() {
+    }, function(err, port) {
+      if (err) return done(err);
+
       pivotServer = spawnServer(`bin/pivot -c test/configs/reintrospect-on-load-datasource.yaml -p ${TEST_PORT}`, {
         env: {
-          DRUID_HOST: 'localhost:28089'
+          DRUID_HOST: `localhost:${port}`
         }
       });
 
@@ -159,6 +162,7 @@ describe('reintrospect on load with datasource', function () {
 
   after(() => {
     pivotServer.kill();
+    druidServer.kill();
   });
 
 });