+++ test/utils/event-collector.js
@@ -14,50 +14,27 @@
  * limitations under the License.
  */
 
-var Q = require('q');
 var http = require('http');
 var express = require('express');
 var bodyParser = require('body-parser');
 
-function mockDruid(options, callback) {
-  var port = options.port || 28082;
-  var version = options.version || '0.9.1';
+function eventCollector(options, callback) {
+  var port = options.port || 28090;
 
   var app = express();
   app.disable('x-powered-by');
 
-  app.get('/status', (req, res) => {
-    if (options.onStatus) {
-      Q(options.onStatus()).then(function(r) {
-        res.status(r.status || 200).json(r.json);
-      });
-    } else {
-      res.json({
-        version
-      });
-    }
-  });
-
-  app.get('/druid/v2/datasources', (req, res) => {
-    if (options.onDataSources) {
-      Q(options.onDataSources(req.body)).then(function(r) {
-        res.status(r.status || 200).json(r.json);
-      });
-    } else {
-      throw new Error('should not GET /druid/v2/datasources');
-    }
-  });
-
   app.use(bodyParser.json());
 
-  app.post('/druid/v2/', (req, res) => {
-    if (options.onQuery) {
-      Q(options.onQuery(req.body)).then(function(r) {
-        res.status(r.status || 200).json(r.json);
-      });
-    } else {
-      throw new Error('should not POST /druid/v2/');
+  var events = [];
+  app.post('/', (req, res) => {
+    if (!Array.isArray(req.body)) {
+      res.status(400).json({ error: 'body must be an array '});
+      return;
     }
+
+    events = events.concat(req.body);
+    res.json({ status: 'ok' });
   });
 
   var server = http.createServer(app);
@@ -80,8 +57,11 @@ function mockDruid(options, callback) {
   return {
     kill: function() {
       server.close();
+    },
+    getEvents: function() {
+      return events;
     }
   };
 }
 
-module.exports = mockDruid;
+module.exports = eventCollector;