+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 For updates follow [@implydata](https://twitter.com/implydata) on Twitter.
 
+## 0.9.28
+
+- Added basic request tracking ability
+- Changed default request log format and made it configurable
+
 ## 0.9.27
 
 - Added gear icon links to settings